Fix build error when imported with SPM
